+++ R/empirical_rxx.R
@@ -7,6 +7,10 @@
 #' @param Theta_SE a matrix of latent trait estimates returned from \code{\link{fscores}} with the options
 #'   \code{full.scores = TRUE} and \code{full.scores.SE = TRUE}
 #'
+#' @param T_as_X logical; should the observed variance be equal to
+#'   \code{var(X) = var(T) + E(E^2)} or \code{var(X) = var(T)} when computing
+#'   empirical reliability estimates? Default (\code{FALSE}) uses the former
+#'
 #' @author Phil Chalmers \email{rphilip.chalmers@@gmail.com}
 #' @references
 #' Chalmers, R., P. (2012). mirt: A Multidimensional Item Response Theory
@@ -20,16 +24,18 @@
 #' \dontrun{
 #'
 #' dat <- expand.table(deAyala)
-#' mod <- mirt(dat, 1)
+#' itemstats(dat)
+#' mod <- mirt(dat)
 #'
 #' theta_se <- fscores(mod, full.scores.SE = TRUE)
 #' empirical_rxx(theta_se)
 #'
 #' theta_se <- fscores(mod, full.scores.SE = TRUE, method = 'ML')
 #' empirical_rxx(theta_se)
+#' empirical_rxx(theta_se, T_as_X = TRUE)
 #'
 #' }
-empirical_rxx <- function(Theta_SE){
+empirical_rxx <- function(Theta_SE, X_as_truescore = FALSE){
     stopifnot(is.matrix(Theta_SE))
     stopifnot(ncol(Theta_SE) %% 2 == 0)
     nms <- colnames(Theta_SE)
@@ -41,6 +47,10 @@ empirical_rxx <- function(Theta_SE){
     pick <- !(rowSums(Theta) %in% c(NA, Inf, -Inf))
     T <- Theta[pick, , drop=FALSE]
     E <- SE[pick, , drop=FALSE]
-    reliability <- diag(var(T)) / (diag(var(T)) + colMeans(E^2))
+    if(T_as_X)
+        reliability <- 1 - colMeans(E^2) / (diag(var(T)))
+    else
+        reliability <- diag(var(T)) / (diag(var(T)) + colMeans(E^2))
+    names(reliability) <- colnames(T)
     reliability
 }